+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## UNRELEASE
 
+- **Add** : Added Apple Pay Multi-browser feature
+
 ## 2.10.0
 
 - **Fix** : fix order not updated due to conflict with WPML plugin